Splitted collection data betweeen service and repository
